requirements.txt
python-dotenv
discord.py==2.6.4
aiohttp